requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
charset-normalizer==3.1.0
discord.py==2.3.1
disnake==2.9.0
distlib==0.3.6
dnspython==2.3.0
filelock==3.12.2
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
pip==23.3.1
platformdirs==3.8.0
pymongo==4.4.0
python-dotenv==1.0.0
setuptools==65.5.0
virtualenv==20.23.1
yarl==1.8.2